Handle infinite values.
Hide text if a ray has infinite length, because otherwise it will show up in the centre. Maybe because translating the canvas by Infinity just does not translate at all.

When drawing a segment, we cannot draw to Infinity, so draw to a large value instead. It will look like a ray to Infinity.